backend/requirements.txt
# RAG Application Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# AI/ML Libraries
google-generativeai==0.4.0
openai==1.12.0
cohere==4.47
pinecone-client==3.0.3
tiktoken==0.6.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0

# For production
gunicorn==21.2.0
